Add adjustable bumper view height
